requirements.txt
dash
dash-bootstrap-components
plotly
pandas
numpy
gunicorn
pyodbc
python-dotenv
